rev 5, fix all problems

+++ src/main/java/com/example/repository/MessageRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.entity.Message;
 
-public interface MessageRepository extends JpaRepository<Message, Long>{
+public interface MessageRepository extends JpaRepository<Message, Integer>{
 
  List<Message> findByPostedBy(int postedBy);
 }
